refactor: Calculation, Validation
- 패키지 이동
- empty.txt 삭제
