+++ src/test/java/techcamp/baseball/CalculationTest.java
@@ -1,8 +1,12 @@
-package techcamp.baseball.domain;
+package techcamp.baseball;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import techcamp.baseball.Calculation;
+import techcamp.baseball.domain.Answer;
+import techcamp.baseball.domain.Question;
+import techcamp.baseball.setting.Config;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
