+++ src/test/java/techcamp/baseball/ValidationTest.java
@@ -1,4 +1,4 @@
-package techcamp.baseball.domain;
+package techcamp.baseball;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class ValidationTest {
